+++ public/js/core/connection.js
@@ -3,9 +3,9 @@
 
     angular.module('app.core').factory('connection', connection);
 
-    connection.$inject = ['$http', 'Noty'];
+    connection.$inject = ['$http', 'notify'];
 
-    function connection ($http, Noty) {
+    function connection ($http, notify) {
         const service = {
             get: get,
             post: post,
@@ -25,9 +25,9 @@
                     if (typeof data === 'string') window.location.href = '/';
 
                     if (data.result === 1 && data.msg && options.showMsg) {
-                        new Noty({ text: data.msg, type: 'success' }).show();
+                        notify.success(data.msg);
                     } else if (data.result === 0 && data.msg && options.showMsg) {
-                        new Noty({ text: data.msg, type: 'error' }).show();
+                        notify.error(data.msg);
                     }
 
                     return data;
@@ -44,9 +44,9 @@
                     if (typeof data === 'string') window.location.href = '/';
 
                     if (data.result === 1 && data.msg) {
-                        new Noty({ text: data.msg, type: 'success' }).show();
+                        notify.success(data.msg);
                     } else if (data.result === 0 && data.msg) {
-                        new Noty({ text: data.msg, type: 'error' }).show();
+                        notify.error(data.msg);
                     }
 
                     return data;