+++ public/js/core/constants.js
@@ -1,4 +1,4 @@
-/* global Noty: false, moment: false, numeral: false */
+/* global PNotify: false, PNotifyBootstrap4: false, moment: false, numeral: false */
 (function () {
     'use strict';
 
@@ -6,7 +6,8 @@
     const base = baseURI.pathname.substring(0, baseURI.pathname.lastIndexOf('/'));
 
     angular.module('app.core')
-        .constant('Noty', Noty)
+        .constant('PNotify', PNotify)
+        .constant('PNotifyBootstrap4', PNotifyBootstrap4)
         .constant('moment', moment)
         .constant('numeral', numeral)
         .constant('base', base);