+++ test/client/about/about.controller.spec.js
@@ -1,6 +1,7 @@
 require('../../../public/js/core/core.module.js');
 require('../../../public/js/core/constants.js');
 require('../../../public/js/core/connection.js');
+require('../../../public/js/core/notification.service');
 require('../../../public/js/core/api.js');
 require('../../../public/js/about/about.module.js');
 require('../../../public/js/about/about.controller.js');