+++ test/client/setup.js
@@ -2,8 +2,10 @@ const jQuery = require('jquery');
 Object.defineProperty(window, 'jQuery', { value: jQuery });
 Object.defineProperty(window, '$', { value: jQuery });
 
-const Noty = require('noty');
-Object.defineProperty(window, 'Noty', { value: Noty });
+const PNotify = require('@pnotify/core');
+const PNotifyBootstrap4 = require('@pnotify/bootstrap4');
+Object.defineProperty(window, 'PNotify', { value: PNotify });
+Object.defineProperty(window, 'PNotifyBootstrap4', { value: PNotifyBootstrap4 });
 
 const moment = require('moment');
 Object.defineProperty(window, 'moment', { value: moment });